third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/event-pseudo-expected.txt

CONSOLE ERROR: line 14: Uncaught TypeError: host.createShadowRoot is not a function
This is a testharness.js-based test.
FAIL event-pseudo Uncaught TypeError: host.createShadowRoot is not a function
Harness: the test ran to completion.
